moved AbstractRecorderConfigurationFactory to org.palladiosimulator.recorderframework.core
