+++ bundles/org.palladiosimulator.recorderframework.core/src/main/java/org/palladiosimulator/recorderframework/core/config/AbstractRecorderConfigurationFactory.java
@@ -1,11 +1,11 @@
-package org.palladiosimulator.recorderframework.config;
+/** */
+package org.palladiosimulator.recorderframework.core.config;
 
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.Map;
 
 import org.palladiosimulator.commons.datastructureutils.MapHelper;
-import org.palladiosimulator.recorderframework.core.config.IRecorderConfigurationFactory;
 
 /**
  * Abstract factory for recorder configuration objects.
