+++ bundles/org.palladiosimulator.recorderframework/META-INF/MANIFEST.MF
@@ -18,7 +18,6 @@ Require-Bundle: org.jscience;bundle-version="4.1.3",
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Export-Package: org.palladiosimulator.recorderframework,
- org.palladiosimulator.recorderframework.config,
  org.palladiosimulator.recorderframework.filters,
  org.palladiosimulator.recorderframework.launch,
  org.palladiosimulator.recorderframework.utils
