--- bundles/org.palladiosimulator.recorderframework/src/org/palladiosimulator/recorderframework/config/package-info.java
@@ -1,6 +1,0 @@
-/**
- * Package for configuration objects of recorders as well as for the creation of such configuration objects.
- * 
- * @author Sebastian Lehrig
- */
-package org.palladiosimulator.recorderframework.config;
